Helper methods for the Moanin class

This is a collection of helper methods for the Moanin class.

# S4 method for Moanin
group_variable_name(object)

# S4 method for Moanin
group_variable_name(object) <- value

# S4 method for Moanin
time_by_group_variable(object)

# S4 method for Moanin
group_variable(object)

# S4 method for Moanin
group_variable(object) <- value

# S4 method for Moanin
time_variable_name(object)

# S4 method for Moanin
time_variable_name(object) <- value

# S4 method for Moanin
time_variable(object)

# S4 method for Moanin
time_variable(object) <- value

# S4 method for Moanin
degrees_of_freedom(object)

# S4 method for Moanin
basis_matrix(object)

# S4 method for Moanin
spline_formula(object)

# S4 method for Moanin
show(object)

# S4 method for Moanin,ANY,character,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for Moanin,ANY,logical,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for Moanin,ANY,numeric,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for Moanin
log_transform(object)

# S4 method for Moanin
get_log_data(object)

Arguments

object
An object of class Moanin, an object containing all related information for time course data and the splines model that will be used (if applicable). See create_moanin_model for more details.

value
replacement value

x
Moanin object

i, j
A vector of logical or integer subscripts, indicating the rows and columns to be subsetted for i and j, respectively.

...
arguments passed to subsetting

drop
A logical scalar that is ignored.

Value

group_variable_name and time_variable_name return the name of the column containing the variable. group_variable and time_variable return the actual variable.

Details

Note that when subsetting the data, the dendrogram information and the co-clustering matrix are lost.

Examples

# Load some data
data(exampleData)
moanin = create_moanin_model(data=testData,meta=testMeta)
group_variable_name(moanin)

#> [1] "Group"

time_variable_name(moanin)

#> [1] "Timepoint"