Create a moanin model

This model is used to conserve information relating to the model, such as the formula, the basis, the meta data.

create_moanin_model(meta, formula = NULL, basis = NULL,
  degrees_of_freedom = NULL)

Arguments

meta
data.frame containing the metadata in columns, and rows corresponding to different samples. Metadata needs to contain the columns "Group" and "Timepoint".

formula
formula object, optional, default: NUlL. Used to construct splines from the data in meta. See details.

basis
matrix, optional, default: NULL. A basis matrix, where each row corresponds to the evaluation of a sample on the basis function (thus one column for each basis function).

degrees_of_freedom
int, optional. Number of degrees of freedom to use if neither the basis nor the formula is provided. If not provided by the user, internally will be set to 4

Value

An object of class moanin_model; a list with the following elements:

basisA basis matrix, where each row corresponds to the evaluation of a sample on the basis function (thus one column for each basis function).

metaThe data frame provided by user to meta

degrees_of_freedomThe degrees of freedom provided by the user (or set by the function internally)

formulaThe formula used to construct the basis matrix (unless basis matrix is provided by the user).

Details

If neither formula nor basis is given, then by default, the function will create a basis matrix based on the formula:

formula = ~Group:ns(Timepoint, df=4) + Group + 0

Examples

# Load some data
data(shoemaker2015)
meta = shoemaker2015$meta

# Use the default options
moanin = create_moanin_model(meta)
print(dim(moanin$basis))

#> [1] 209  25

# Change the number of degrees of freedom
moanin = create_moanin_model(meta, degrees_of_freedom=6)
print(dim(moanin$basis))

#> [1] 209  35