Performs splines clustering using K-means

Performs splines clustering using K-means

splines_kmeans(data, moanin_model, n_clusters = 10, init = "kmeans++",
  n_init = 10, max_iter = 300, random_seed = NULL,
  fit_splines = TRUE, rescale = TRUE)

Arguments

data
The data matrix, where genes (features) are on the rows and samples on the columns.

moanin_model
object of class moanin_model, an object containing all related information for time course data and the splines model that will be used (if applicable). See create_moanin_model for more details.

n_clusters
int optional, default: 10

init
["kmeans++", "random", "optimal_init"]

n_init
int, optional, default: 10 Number of initialization to perform.

max_iter
int, optional, default: 300 Maximum number of iteration to perform

random_seed
int, optional, default: NULL

fit_splines
boolean, optional, default: TRUE Whether to fit splines or not.

rescale
boolean, optional, default: TRUE Whether to rescale the data or not. @return A list in the format returned by KMeans_rcpp, with the following elements added or changed:

centroidsThe centroids are rescaled so that they range from 0-1

moanin_modelThe given moanin_model

fit_splinesThe value of fit_splines given to the function

rescaleThe value of rescale given to the function

Examples

data(exampleData)
# Use the default options
moanin = create_moanin_model(testMeta)
splines_kmeans(testData, moanin)

#> $clusters
#>          NM_009912          NM_008725          NM_007473 ENSMUST00000094955 
#>                  3                  4                 10                  4 
#>       NM_001042489          NM_008159       NM_001013813           AK039774 
#>                  5                  3                  1                  5 
#>          NM_013782          NM_028622          NM_028007          NM_198093 
#>                  1                  9                  5                  6 
#>          NM_145123          NM_027869          NM_011448          NM_027495 
#>                  1                  3                  2                  1 
#>          NM_153399          NM_031158          NM_008394          NM_173395 
#>                  4                  1                  3                  8 
#>       NM_001033490          NM_182991          NM_010700       NM_001005247 
#>                 10                  7                  5                  4 
#>           AK042548          NM_198703           AK032795          NM_009127 
#>                  8                  9                  9                  5 
#>          NM_145418           BG963265          NM_172503          NM_023333 
#>                 10                  1                  2                  3 
#>        NAP103812.1          NM_182990          NM_145555          NM_025780 
#>                  2                  8                  8                  7 
#>           BC003314          NM_026232          NM_028234             X59251 
#>                  8                  2                  2                  5 
#>          NM_146235          NM_027604          NM_026039          NM_176958 
#>                  8                  8                  2                  1 
#>          NM_032393          NM_026779          NM_172520          NM_146073 
#>                  8                  3                  1                  5 
#>           AK044468          NM_010397          NM_139148          NM_080846 
#>                  2                  3                  3                  4 
#>          NM_153389          NM_029920          NM_009448          NM_026890 
#>                  3                  6                  2                  3 
#>          NM_031189          NM_139065           AK148746           AK129092 
#>                  3                  2                  7                  9 
#>           AK013290           AK077042          NM_022886          NM_011895 
#>                  9                  1                  5                  1 
#>          NM_025521          NM_172785          NM_022995          NM_146308 
#>                  2                  8                 10                  9 
#>          NM_172590          NM_146261          NM_026854          NM_022325 
#>                  1                  4                  4                  8 
#>          NM_178682       NM_001042719          NM_172730          NM_021388 
#>                  3                  2                 10                  2 
#>           AK077979          NM_011358          NR_015493 ENSMUST00000019066 
#>                  5                  2                  1                  8 
#>          NM_175238           AK080090          NM_021487           AK048544 
#>                  6                  8                 10                  2 
#>          NM_177821           AK050809          NM_172118 ENSMUST00000110434 
#>                  8                  2                  1                  9 
#>             M60510          NM_009046          NM_175251          NM_007570 
#>                  9                  3                  5                  4 
#>          NM_009450          NM_026778          NM_030143          NM_080448 
#>                 10                  5                  1                  7 
#>          NM_011543          NM_007790          NM_028002          NM_030097 
#>                  6                  1                 10                  5 
#>          NM_010806           AK020388       A_52_P692753       NM_001081274 
#>                  6                  1                 10                  2 
#>          NM_172522           BC106120           AK003606           AK077928 
#>                  3                  5                 10                  6 
#>          NM_011756          NM_029658             M29009          NM_172648 
#>                  2                  4                  7                  3 
#>          NM_010662          NM_030265          NM_010749        NAP030421.1 
#>                  8                  5                  7                  9 
#>          NM_010705          NM_010636       NM_001081286          NM_181750 
#>                  8                  1                  1                  2 
#>          NM_007789           AK051283           AK049456          NM_011437 
#>                  4                  5                  1                  9 
#>          NM_009096          NM_009573          NM_201363          NM_030057 
#>                  6                  3                  9                  4 
#>          NM_009148           AK084873          NM_007866          NM_028932 
#>                  1                  5                 10                  7 
#>          NM_138309          NM_173370          NM_010000           AK079518 
#>                  1                  8                 10                  8 
#>          NM_008796          NM_016742           AK050890          NM_177855 
#>                  2                  1                  9                  3 
#>          NM_173368       NM_001081557      NAP014233.001        NAP034216.1 
#>                  5                  6                  3                 10 
#>          NM_147073          NM_013886       NM_001081324           AK086626 
#>                  9                  6                  1                  4 
#>          NM_010217          NM_016712          NM_028270             U49929 
#>                  8                 10                  3                  8 
#>          NM_008298          NM_009642          NM_010452          NM_009864 
#>                  7                  8                  1                  1 
#>           AK140034       NM_001081270          NM_011303          NM_019782 
#>                  5                  3                  5                  2 
#>        NAP113251.1           AV264460       NM_001025566          NM_019639 
#>                  1                  1                  5                 10 
#> ENSMUST00000105872           AK078573          NM_007956           AK038771 
#>                  1                  5                  8                  6 
#>          NM_028244          NM_016859          NM_010927          NM_144943 
#>                  9                  3                  9                  5 
#>           BC013561           AK050631          NM_013680          NM_026247 
#>                  4                  9                  9                  8 
#>          NM_011348 ENSMUST00000028726          NM_027485          NM_026721 
#>                  5                  5                  7                  1 
#>          NM_025892          NM_145615          NM_026191           AK003805 
#>                  5                  1                 10                  9 
#>          NM_138679          NM_172478          NM_009748          NM_009829 
#>                  4                 10                 10                  3 
#>       NM_001081651       NM_001039094 ENSMUST00000062252          NM_009011 
#>                  2                  5                  9                  7 
#>          NM_022026          NM_013551          NM_027264       NM_001037846 
#>                  9                  8                 10                  4 
#>       NM_001013577       NM_001081644          XM_906012          NM_007436 
#>                  3                  4                  8                  5 
#>           AK080901          NM_023061             X89687          NM_080850 
#>                  1                 10                  9                  8 
#>          NM_023608          NM_054039           AK045942          NM_145216 
#>                 10                  8                  8                  5 
#>          NM_013827          NM_177912             X14206       NM_001081012 
#>                  1                  9                  2                  6 
#>          NM_027263          NM_172715          NM_183097          NM_025299 
#>                  8                  4                  9                  2 
#>          NM_010205       NM_001081282          NM_177305       NM_001017983 
#>                  7                  7                  2                  7 
#>          NM_027250       NM_001093750           AB070552          NM_008745 
#>                  8                  9                  2                  4 
#>           AK089219          NM_009087          NM_028319          NM_019442 
#>                  3                  9                  5                  2 
#>          NM_011317          NM_011581           AK042550          NM_029454 
#>                  4                  8                  5                  4 
#>           AK036657          NM_146531          NM_011288       NM_001081113 
#>                  6                  2                  1                  6 
#>           AK053721           AK053770        NAP045501.1           AK076230 
#>                 10                  3                  3                  7 
#>       NM_001033532          NM_010934          NM_013925        NM_027495.1 
#>                  4                  5                  3                  1 
#>        NAP052665.1        NAP032196.1          NM_010852          NM_146137 
#>                 10                  7                  8                  1 
#>          NM_013732          NM_008435          NM_011363 ENSMUST00000092720 
#>                 10                  9                  1                  7 
#>          NM_027172          NM_054102          NM_145467           AK133434 
#>                  5                  1                  4                  5 
#>          NM_145548          NM_053084          NM_172704          NM_133192 
#>                 10                  1                 10                  5 
#>          NM_010773          NM_145456          NM_011836          NM_025794 
#>                  2                  6                 10                  6 
#> ENSMUST00000079389          NM_007554       A_51_P498312       NM_001145300 
#>                  1                  1                  3                  9 
#>          NM_178069          NM_016868          NM_177473       A_52_P764477 
#>                  3                  3                  4                  7 
#>          NM_010957          NM_010704          NM_010544       NM_001018042 
#>                  4                  5                  2                  1 
#>          NM_172815           AK077058          NM_025899           BC002117 
#>                  5                  2                  4                  4 
#>           AK049625          NM_178764           AK086513          NM_153117 
#>                  5                  1                  4                  3 
#>           AK080070           AK035035          NM_010548          NM_010446 
#>                  5                  9                  8                  5 
#>          NM_009710           BQ033444          NM_029702          NM_194060 
#>                  4                  3                  9                  5 
#>           BC025043          NM_201530          XM_983040          NM_133750 
#>                  8                  9                  9                  9 
#>          NM_009135        NAP057119.1        NM_010548.1          NM_029803 
#>                  5                  3                  8                  3 
#>          NM_207280       NM_001081078           AK031104          NM_025333 
#>                  5                  2                  1                  8 
#>        NAP067097.1           AK051337           AK030165          NM_153100 
#>                  3                  5                  2                  5 
#>          NM_175226          NM_026149          NM_177868           AK028908 
#>                  3                  8                  5                  7 
#>       NM_001081406             S76748 ENSMUST00000039333          NM_009875 
#>                  8                  2                  6                  5 
#>       NM_001005370          NM_015747          NM_026360          NM_031196 
#>                  4                  7                  2                  9 
#>           AK051796          NM_025543          NM_025952       NM_001013770 
#>                  6                  6                  8                  2 
#>          NM_147054          NM_019564           AB026045          NM_009735 
#>                  9                  1                  3                  3 
#>       NM_001033431       NM_001029838          NM_011983          NM_010442 
#>                  2                  2                  1                  3 
#>       NM_001008533          NM_022891          NM_207275           AK050046 
#>                  3                  4                  9                  7 
#>          NM_146182           AK021064       XM_001475276          XM_126677 
#>                  1                  8                  5                  9 
#>          NM_172663          NM_018788        NAP029329.1           U49929.1 
#>                  5                  6                  1                  8 
#>           AK032619       NM_001001566 ENSMUST00000050234          NM_183106 
#>                  4                  1                  9                  2 
#>           AF223362       NM_001081079          NM_025816       NM_001029937 
#>                  9                  2                  6                  4 
#>          NM_178076          NM_032418          NM_028673           AF012177 
#>                  1                  9                  9                  8 
#>          NM_178257       NM_001038845          NM_019764       NM_001037127 
#>                  9                  8                 10                  9 
#>          NM_010684        NM_009642.1           AK038621          NM_201646 
#>                  5                  7                  9                  1 
#>          NM_177678 ENSMUST00000103303          NM_007776          NM_172691 
#>                  5                  1                  3                  2 
#>           AK078805          NM_198030        NAP061200.1 ENSMUST00000063103 
#>                  4                  3                  3                  8 
#>        NAP031516.1       NM_001039146          NM_011378          NM_175026 
#>                  2                  5                  1                  3 
#>           AK083006          NM_023852          NM_174875          NM_028181 
#>                  1                  1                  7                  1 
#>       XM_001478619          NM_030024          NM_028982       NM_001025572 
#>                  6                  2                  1                  4 
#>       NM_001033599           AK035513          NM_172553           AK050536 
#>                  8                  7                  8                  5 
#>          NM_133789 ENSMUST00000024982           AK089906           AK033780 
#>                 10                  5                  3                  6 
#>           AK086311          NM_026458          NM_021392          NM_025611 
#>                  5                  9                  1                  1 
#>          NM_009539          NM_011779          NM_009941          NM_007567 
#>                  8                  7                  4                  3 
#>          NM_009572          NM_153601          NM_013836          NM_025813 
#>                  2                  9                  9                  5 
#>           AK041701           AK049540          NM_010593          NM_008218 
#>                  5                  9                  1                  5 
#>          XM_891200          NM_013527          NM_011261          NM_010251 
#>                  5                  7                  5                  4 
#>       XM_001477170       NM_001081152          NM_177583           AB056117 
#>                  2                  6                  4                  8 
#>       NM_001003972          NM_146949          NM_177033          NM_026933 
#>                  6                  6                  1                  1 
#>          NM_007476           AK016752           AK080095           AK083360 
#>                 10                  5                  8                  8 
#>       NM_001081223          NR_015599          NM_009325          NM_170759 
#>                  8                  7                  1                  9 
#>          NM_183225          NM_033567          NM_010145          NM_145404 
#>                  8                  5                  5                  8 
#>       NM_001039552          NM_009007          NM_028487          NM_008914 
#>                  2                  8                  2                  6 
#>          NM_027453           AK031823          NM_021477          NM_024257 
#>                  5                  4                  1                  1 
#>          NM_013721          NM_028443        NM_138679.1          NM_008312 
#>                  1                  5                  1                  9 
#>          NM_031168          NM_172269           AK140681           AK053751 
#>                  3                  6                  5                  9 
#>          NM_009808          NM_008382          NM_008602          NM_013599 
#>                  3                  9                  1                 10 
#>          NM_173780           AK048250       NM_001081290          NM_027118 
#>                  9                  4                  7                  4 
#>          NM_026253           AK031225          NM_178674       NM_001033481 
#>                  5                  5                  4                  1 
#>           AK011948          NM_133900          NM_146443          NM_025905 
#>                  4                  7                  3                  1 
#>      NAP020355.001 ENSMUST00000106746           AK052576          NM_021389 
#>                  5                  4                  6                  8 
#>          NM_011763          NM_008991          NM_010257          NM_030261 
#>                  6                  5                  3                  8 
#>           AK019056          NM_019432       NM_001024539           AK035407 
#>                  4                  2                  3                  7 
#>           AK017896       NM_001033270           BC004066          NM_007930 
#>                  5                  8                  9                  3 
#>          NM_144846          NM_147068          NM_178906          NM_012000 
#>                  8                  9                  5                  6 
#>          NM_011577          NM_172736          NM_008121           AK087086 
#>                  2                  1                  9                 10 
#>          NM_177854          NM_010794          NM_008950          NM_172806 
#>                  9                  9                  3                 10 
#>           AK016162           AK034785       NM_001010937          NM_029680 
#>                  9                  7                  5                  9 
#>          NM_007691           AK031401       NM_001163774          NM_026944 
#>                  8                  7                  5                  7 
#> 
#> $centroids
#>     GSM1557140 GSM1557141 GSM1557142  GSM1557143  GSM1557144  GSM1557145
#> C1  0.89121698 0.89121699 0.89121698 0.991055823 0.991055823 0.991055823
#> C2  0.65928018 0.65928018 0.65928018 0.839534963 0.839534963 0.839534963
#> C3  0.23943543 0.23943543 0.23943543 0.154713756 0.154713756 0.154713756
#> C4  0.56132658 0.56132658 0.56132658 0.543902547 0.543902547 0.543902547
#> C5  0.89211686 0.89211686 0.89211686 0.947139949 0.947139949 0.947139949
#> C6  0.22549122 0.22549122 0.22549122 0.131834288 0.131834288 0.131834288
#> C7  0.15923650 0.15923650 0.15923650 0.005559943 0.005559943 0.005559943
#> C8  0.04679047 0.04679047 0.04679047 0.073400398 0.073400398 0.073400398
#> C9  0.76877844 0.76877844 0.76877844 0.906252781 0.906252781 0.906252781
#> C10 0.84920379 0.84920379 0.84920379 1.000000000 1.000000000 1.000000000
#>     GSM1557146 GSM1557147 GSM1557148 GSM1557149 GSM1557150 GSM1557151
#> C1   0.9065147  0.9065147  0.9065147 0.77629894 0.77629894 0.77629894
#> C2   0.7612802  0.7612802  0.7612802 0.62361803 0.62361803 0.62361803
#> C3   0.2576678  0.2576678  0.2576678 0.39889123 0.39889123 0.39889123
#> C4   0.4871197  0.4871197  0.4871197 0.42445550 0.42445550 0.42445550
#> C5   0.9129064  0.9129064  0.9129064 0.85906059 0.85906059 0.85906059
#> C6   0.1903939  0.1903939  0.1903939 0.29151843 0.29151843 0.29151843
#> C7   0.1030763  0.1030763  0.1030763 0.25964049 0.25964049 0.25964049
#> C8   0.0692015  0.0692015  0.0692015 0.05975563 0.05975563 0.05975563
#> C9   0.9604045  0.9604045  0.9604045 0.99431830 0.99431830 0.99431830
#> C10  0.9560275  0.9560275  0.9560275 0.86540578 0.86540578 0.86540578
#>     GSM1557152 GSM1557153 GSM1557154 GSM1557155 GSM1557156 GSM1557157
#> C1  0.65105339 0.65105339 0.65105339  0.5540552  0.5540552  0.5540552
#> C2  0.48801300 0.48801300 0.48801300  0.3830613  0.3830613  0.3830613
#> C3  0.51386900 0.51386900 0.51386900  0.5756207  0.5756207  0.5756207
#> C4  0.38323463 0.38323463 0.38323463  0.3736257  0.3736257  0.3736257
#> C5  0.81005750 0.81005750 0.81005750  0.7767485  0.7767485  0.7767485
#> C6  0.40147303 0.40147303 0.40147303  0.5029673  0.5029673  0.5029673
#> C7  0.40942704 0.40942704 0.40942704  0.5222934  0.5222934  0.5222934
#> C8  0.05227126 0.05227126 0.05227126  0.0498759  0.0498759  0.0498759
#> C9  1.00000000 1.00000000 1.00000000  0.9772195  0.9772195  0.9772195
#> C10 0.76770194 0.76770194 0.76770194  0.6822366  0.6822366  0.6822366
#>     GSM1557158 GSM1557159 GSM1557160 GSM1557161 GSM1557162 GSM1557163
#> C1  0.94119209 0.94119209 0.94119209 0.42724906 0.42724906 0.42724906
#> C2  0.73551332 0.73551332 0.73551332 0.24659947 0.24659947 0.24659947
#> C3  0.19043320 0.19043320 0.19043320 0.58329431 0.58329431 0.58329431
#> C4  0.56665511 0.56665511 0.56665511 0.42054835 0.42054835 0.42054835
#> C5  0.91745649 0.91745649 0.91745649 0.74561285 0.74561285 0.74561285
#> C6  0.18295037 0.18295037 0.18295037 0.67790633 0.67790633 0.67790633
#> C7  0.08836731 0.08836731 0.08836731 0.66097861 0.66097861 0.66097861
#> C8  0.05674901 0.05674901 0.05674901 0.05636097 0.05636097 0.05636097
#> C9  0.80477611 0.80477611 0.80477611 0.87674598 0.87674598 0.87674598
#> C10 0.90867800 0.90867800 0.90867800 0.54343640 0.54343640 0.54343640
#>     GSM1557164 GSM1557165 GSM1557166 GSM1557167 GSM1557168 GSM1557169
#> C1  0.35482673 0.35482673 0.35482673 0.98088161 0.98088161 0.98088161
#> C2  0.17005001 0.17005001 0.17005001 0.79893411 0.79893411 0.79893411
#> C3  0.51776912 0.51776912 0.51776912 0.15288846 0.15288846 0.15288846
#> C4  0.50410256 0.50410256 0.50410256 0.56798469 0.56798469 0.56798469
#> C5  0.73949660 0.73949660 0.73949660 0.93803098 0.93803098 0.93803098
#> C6  0.81457299 0.81457299 0.81457299 0.14826283 0.14826283 0.14826283
#> C7  0.73087935 0.73087935 0.73087935 0.03084088 0.03084088 0.03084088
#> C8  0.07034129 0.07034129 0.07034129 0.06534706 0.06534706 0.06534706
#> C9  0.75553300 0.75553300 0.75553300 0.84012876 0.84012876 0.84012876
#> C10 0.43543406 0.43543406 0.43543406 0.95920882 0.95920882 0.95920882
#>     GSM1557170 GSM1557171 GSM1557172   GSM1557173 GSM1557174   GSM1557175
#> C1  0.30136047 0.30136047 0.30136047 1.000000e+00 1.00000000 1.000000e+00
#> C2  0.11513852 0.11513852 0.11513852 8.367302e-01 0.83673021 8.367302e-01
#> C3  0.45983600 0.45983600 0.45983600 1.382587e-01 0.13825869 1.382587e-01
#> C4  0.57320016 0.57320016 0.57320016 5.613164e-01 0.56131638 5.613164e-01
#> C5  0.73614136 0.73614136 0.73614136 9.490752e-01 0.94907518 9.490752e-01
#> C6  0.91282256 0.91282256 0.91282256 1.292819e-01 0.12928190 1.292819e-01
#> C7  0.77949767 0.77949767 0.77949767 1.267591e-16 0.00000000 6.337953e-17
#> C8  0.08430639 0.08430639 0.08430639 7.122414e-02 0.07122414 7.122414e-02
#> C9  0.66527598 0.66527598 0.66527598 8.741914e-01 0.87419135 8.741914e-01
#> C10 0.34588393 0.34588393 0.34588393 9.918529e-01 0.99185288 9.918529e-01
#>      GSM1557176  GSM1557177  GSM1557178   GSM1557179   GSM1557180   GSM1557181
#> C1  0.181919926 0.181919926 0.181919926 1.488103e-01 1.488103e-01 1.488103e-01
#> C2  0.006721666 0.006721666 0.006721666 8.374386e-17 4.187193e-17 0.000000e+00
#> C3  0.417026058 0.417026058 0.417026058 5.469453e-01 5.469453e-01 5.469453e-01
#> C4  0.653317976 0.653317976 0.653317976 5.539866e-01 5.539866e-01 5.539866e-01
#> C5  0.709271841 0.709271841 0.709271841 6.701065e-01 6.701065e-01 6.701065e-01
#> C6  1.000000000 1.000000000 1.000000000 8.075698e-01 8.075698e-01 8.075698e-01
#> C7  0.880250390 0.880250390 0.880250390 8.953157e-01 8.953157e-01 8.953157e-01
#> C8  0.128097449 0.128097449 0.128097449 1.608536e-01 1.608536e-01 1.608536e-01
#> C9  0.632731125 0.632731125 0.632731125 9.005138e-01 9.005138e-01 9.005138e-01
#> C10 0.114821639 0.114821639 0.114821639 0.000000e+00 8.532497e-17 4.266248e-17
#>     GSM1557224 GSM1557225 GSM1557226   GSM1557227   GSM1557228   GSM1557229
#> C1  0.46837592 0.46837592 0.46837592 8.568900e-01 8.568900e-01 8.568900e-01
#> C2  0.28286993 0.28286993 0.28286993 5.160267e-01 5.160267e-01 5.160267e-01
#> C3  0.21952213 0.21952213 0.21952213 3.317615e-03 3.317615e-03 3.317615e-03
#> C4  0.91451497 0.91451497 0.91451497 9.678900e-01 9.678900e-01 9.678900e-01
#> C5  0.88134823 0.88134823 0.88134823 1.000000e+00 1.000000e+00 1.000000e+00
#> C6  0.77024385 0.77024385 0.77024385 4.292166e-01 4.292166e-01 4.292166e-01
#> C7  0.91116851 0.91116851 0.91116851 8.563872e-01 8.563872e-01 8.563872e-01
#> C8  0.02223598 0.02223598 0.02223598 0.000000e+00 6.989982e-17 6.989982e-17
#> C9  0.33559308 0.33559308 0.33559308 1.925065e-17 0.000000e+00 3.850130e-17
#> C10 0.41960732 0.41960732 0.41960732 7.682468e-01 7.682468e-01 7.682468e-01
#>      GSM1557230  GSM1557231  GSM1557232 GSM1557233 GSM1557234 GSM1557235
#> C1  0.791435580 0.791435580 0.791435580 0.62882855 0.62882855 0.62882855
#> C2  0.620908279 0.620908279 0.620908279 0.70349097 0.70349097 0.70349097
#> C3  0.123376779 0.123376779 0.123376779 0.32274171 0.32274171 0.32274171
#> C4  0.816210510 0.816210510 0.816210510 0.60928276 0.60928276 0.60928276
#> C5  0.963958074 0.963958074 0.963958074 0.88337197 0.88337197 0.88337197
#> C6  0.262094861 0.262094861 0.262094861 0.12281555 0.12281555 0.12281555
#> C7  0.823883738 0.823883738 0.823883738 0.79879871 0.79879871 0.79879871
#> C8  0.004426031 0.004426031 0.004426031 0.01883262 0.01883262 0.01883262
#> C9  0.095416579 0.095416579 0.095416579 0.27808775 0.27808775 0.27808775
#> C10 0.851924983 0.851924983 0.851924983 0.88670771 0.88670771 0.88670771
#>     GSM1557236 GSM1557237 GSM1557238   GSM1557239   GSM1557240 GSM1557241
#> C1  0.45899670 0.45899670 0.45899670 3.250279e-01 3.250279e-01 0.32502790
#> C2  0.75199640 0.75199640 0.75199640 7.648723e-01 7.648723e-01 0.76487229
#> C3  0.51705479 0.51705479 0.51705479 6.668952e-01 6.668952e-01 0.66689520
#> C4  0.41902174 0.41902174 0.41902174 2.777276e-01 2.777276e-01 0.27772756
#> C5  0.78345090 0.78345090 0.78345090 6.787589e-01 6.787589e-01 0.67875894
#> C6  0.03369867 0.03369867 0.03369867 7.372873e-17 3.686436e-17 0.00000000
#> C7  0.79113248 0.79113248 0.79113248 8.033029e-01 8.033029e-01 0.80330287
#> C8  0.04152634 0.04152634 0.04152634 7.097771e-02 7.097771e-02 0.07097771
#> C9  0.44698873 0.44698873 0.44698873 5.565880e-01 5.565880e-01 0.55658800
#> C10 0.89071200 0.89071200 0.89071200 8.752829e-01 8.752829e-01 0.87528289
#>     GSM1557242 GSM1557243 GSM1557244 GSM1557245 GSM1557246 GSM1557247
#> C1  0.61318812 0.61318812 0.61318812 0.14763465 0.14763465 0.14763465
#> C2  0.34203494 0.34203494 0.34203494 0.71887536 0.71887536 0.71887536
#> C3  0.12315111 0.12315111 0.12315111 0.85717561 0.85717561 0.85717561
#> C4  0.96142157 0.96142157 0.96142157 0.11105847 0.11105847 0.11105847
#> C5  0.92789164 0.92789164 0.92789164 0.47027513 0.47027513 0.47027513
#> C6  0.68624981 0.68624981 0.68624981 0.05088178 0.05088178 0.05088178
#> C7  0.89907397 0.89907397 0.89907397 0.86616510 0.86616510 0.86616510
#> C8  0.01411865 0.01411865 0.01411865 0.14469849 0.14469849 0.14469849
#> C9  0.20257298 0.20257298 0.20257298 0.64014127 0.64014127 0.64014127
#> C10 0.52234575 0.52234575 0.52234575 0.80412257 0.80412257 0.80412257
#>     GSM1557248 GSM1557249 GSM1557250 GSM1557251 GSM1557252 GSM1557253
#> C1  0.05237668 0.05237668 0.05237668 0.73840872 0.73840872 0.73840872
#> C2  0.63800300 0.63800300 0.63800300 0.40086958 0.40086958 0.40086958
#> C3  0.95138700 0.95138700 0.95138700 0.04417782 0.04417782 0.04417782
#> C4  0.03870098 0.03870098 0.03870098 0.99451947 0.99451947 0.99451947
#> C5  0.28603196 0.28603196 0.28603196 0.96751403 0.96751403 0.96751403
#> C6  0.17580032 0.17580032 0.17580032 0.60169325 0.60169325 0.60169325
#> C7  0.94376013 0.94376013 0.94376013 0.88630637 0.88630637 0.88630637
#> C8  0.22912067 0.22912067 0.22912067 0.00704560 0.00704560 0.00704560
#> C9  0.62533100 0.62533100 0.62533100 0.08975379 0.08975379 0.08975379
#> C10 0.70775072 0.70775072 0.70775072 0.61870813 0.61870813 0.61870813
#>       GSM1557254   GSM1557255 GSM1557256   GSM1557257   GSM1557258  GSM1557259
#> C1  7.220759e-17 3.610380e-17  0.0000000 8.244462e-01 8.244462e-01 0.824446154
#> C2  5.823397e-01 5.823397e-01  0.5823397 4.590435e-01 4.590435e-01 0.459043477
#> C3  1.000000e+00 1.000000e+00  1.0000000 8.435379e-17 1.687076e-17 0.000000000
#> C4  0.000000e+00 0.000000e+00  0.0000000 1.000000e+00 1.000000e+00 1.000000000
#> C5  1.481090e-01 1.481090e-01  0.1481090 9.932944e-01 9.932944e-01 0.993294379
#> C6  2.915751e-01 2.915751e-01  0.2915751 5.160116e-01 5.160116e-01 0.516011634
#> C7  1.000000e+00 1.000000e+00  1.0000000 8.721927e-01 8.721927e-01 0.872192665
#> C8  3.151914e-01 3.151914e-01  0.3151914 2.061115e-03 2.061115e-03 0.002061115
#> C9  5.946423e-01 5.946423e-01  0.5946423 1.733641e-02 1.733641e-02 0.017336414
#> C10 6.145250e-01 6.145250e-01  0.6145250 7.023184e-01 7.023184e-01 0.702318429
#>     GSM1557260 GSM1557261   GSM1557262 GSM1557263 GSM1557264 GSM1557265
#> C1  0.06914484 0.06914484 6.914484e-02 0.39305685 0.39305685 0.39305685
#> C2  0.65580261 0.65580261 6.558026e-01 1.00000000 1.00000000 1.00000000
#> C3  0.91203167 0.91203167 9.120317e-01 0.56585046 0.56585046 0.56585046
#> C4  0.01642157 0.01642157 1.642157e-02 0.18939216 0.18939216 0.18939216
#> C5  0.00000000 0.00000000 1.680650e-17 0.22088103 0.22088103 0.22088103
#> C6  0.54960740 0.54960740 5.496074e-01 0.62014995 0.62014995 0.62014995
#> C7  0.99767616 0.99767616 9.976762e-01 0.78755042 0.78755042 0.78755042
#> C8  0.65818500 0.65818500 6.581850e-01 1.00000000 1.00000000 1.00000000
#> C9  0.47680801 0.47680801 4.768080e-01 0.36347253 0.36347253 0.36347253
#> C10 0.31176209 0.31176209 3.117621e-01 0.07312735 0.07312735 0.07312735
#> 
#> $total_SSE
#> [1] 3728.161
#> 
#> $best_initialization
#> [1] 9
#> 
#> $WCSS_per_cluster
#>          [,1]     [,2]     [,3]    [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 159.4613 162.5593 138.4665 133.627 135.8148 90.55374 100.0656 117.0877
#>          [,9]    [,10]
#> [1,] 201.2661 90.42181
#> 
#> $obs_per_cluster
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]   70   48   53   44   73   30   32   58   59    33
#> 
#> $between.SS_DIV_total.SS
#> [1] 0.6434371
#> 
#> $moanin_model
#> moanin_model object on 84 samples containing the following information:
#> 1) Meta data with 5 variables
#> [1] ""            "Group"       "Replicate"   "Timepoint"   "WeeklyGroup"
#> 2) Group variable given by 'Group'
#> 3) Time variable given by 'Timepoint'
#> 4) Basis matrix with 25 basis functions
#> Basis matrix was constructed with the following formula
#> ~Group + Group:splines::ns(Timepoint, df = 4) + 0 
#> To access this information use:
#> 	<object_name>$degrees_of_freedom
#> 	<object_name>$basis
#> 	<object_name>$meta
#> 	<object_name>$formula
#> 	<object_name>$time_variable
#> 	<object_name>$group_variable
#> $fit_splines
#> [1] TRUE
#> 
#> $rescale
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "k-means clustering"