Assign score and labels from raw data

Assign score and labels from raw data

splines_kmeans_score_and_label(data, kmeans_clusters,
  percentage_genes_to_label = 0.5, max_score = NULL,
  rescale_separately_on = NULL)

Arguments

data
The data matrix, where genes (features) are on the rows and samples on the columns.

kmeans_clusters
list of list List returned by moanin::splines_kmeans

percentage_genes_to_label
float, optional, default: 0.5 Percentage of genes to label. If max_score is provided, will label genes that are either in the top `percentage_genes_to_label` or with a score below `max_score`.

max_score
optional, default: Null When provided, will only label genes below that score. If NULL, ignore this option.

rescale_separately_on,
string, optional, default: NULL When provided, will rescale separately different groups of data. @return A list consisting of

labelsthe label or cluster assigned to each gene based on the cluster with the best (i.e. lowest) score, with no label given to genes that do not have a score lower than a specified quantity

scoresthe matrix of size n_cluster x n_genes, containing for each gene and each cluster, the goodness of fit score

max_scoreThe required cutoff for a gene receiving an assignment